refactor: move pages to subdirectories
