+++ taxonomy-editor-frontend/src/pages/project/editentry/ListAllEntryProperties.tsx
